+++ finished/02_libft/ft_memccpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memccpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mrhea-ro <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: mrhea-ro <mrhea-ro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/24 22:28:31 by mrhea-ro          #+#    #+#             */
-/*   Updated: 2018/12/25 19:14:26 by mrhea-ro         ###   ########.fr       */
+/*   Updated: 2019/05/18 18:31:46 by mrhea-ro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 {
-	size_t			i;
+	int				i;
 	unsigned char	*source;
 	unsigned char	*dest;
 
 	source = (unsigned char *)src;
 	dest = (unsigned char *)dst;
 	i = -1;
-	while (++i < n)
+	while (++i < (int)n)
 	{
 		dest[i] = source[i];
 		if ((char)source[i] == (char)c)
