+++ finished/02_libft/ft_memchr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mrhea-ro <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: mrhea-ro <mrhea-ro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/24 10:45:32 by mrhea-ro          #+#    #+#             */
-/*   Updated: 2018/12/25 19:14:06 by mrhea-ro         ###   ########.fr       */
+/*   Updated: 2019/05/19 19:31:21 by mrhea-ro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	size_t		i;
+	int			i;
 	const char	*buff4ik;
 
 	buff4ik = s;
 	i = -1;
-	while (++i < n)
+	while (++i < (int)n)
 		if (buff4ik[i] == (char)c)
 			return ((void *)(s + i));
 	return (NULL);
